backend/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.10.0
psycopg2-binary
Pillow